+++ albam/engines/mtfw/archive.py
@@ -1,129 +1,39 @@
 import io
 import ntpath
-import os
-from pathlib import PureWindowsPath, Path
 import zlib
 
 from kaitaistruct import KaitaiStream
 
 from ...registry import blender_registry
+from ...lib.kaitai_utils import check_recursive
 from . import EXTENSION_TO_FILE_ID, FILE_ID_TO_EXTENSION
+from .arc_fs import ArcFS, MTFW_FS
 from .structs.arc import Arc
 from ...blender_ui.tools_panel import show_message_box
 
 
-@blender_registry.register_archive_loader(app_id="re0", extension="arc")
-@blender_registry.register_archive_loader(app_id="re1", extension="arc")
-@blender_registry.register_archive_loader(app_id="re5", extension="arc")
-@blender_registry.register_archive_loader(app_id="re6", extension="arc")
-@blender_registry.register_archive_loader(app_id="rev1", extension="arc")
-@blender_registry.register_archive_loader(app_id="rev2", extension="arc")
-@blender_registry.register_archive_loader(app_id="dd", extension="arc")
-@blender_registry.register_archive_loader(app_id="dmc4", extension="arc")
-def arc_loader(vfile, context=None):  # XXX context DEPRECATED
-    arc = ArcWrapper(file_path=vfile.absolute_path)
-    for file_entry in arc.get_file_entries():
-        yield file_entry.file_path_with_ext
+@blender_registry.register_fs_root_loader(app_id="re0", extension="arc")
+@blender_registry.register_fs_root_loader(app_id="re1", extension="arc")
+@blender_registry.register_fs_root_loader(app_id="re5", extension="arc")
+@blender_registry.register_fs_root_loader(app_id="re6", extension="arc")
+@blender_registry.register_fs_root_loader(app_id="rev1", extension="arc")
+@blender_registry.register_fs_root_loader(app_id="rev2", extension="arc")
+@blender_registry.register_fs_root_loader(app_id="dd", extension="arc")
+@blender_registry.register_fs_root_loader(app_id="dmc4", extension="arc")
+def arc_fs_root_loader(absolute_path):
+    return ArcFS(absolute_path)
 
 
-@blender_registry.register_archive_accessor(app_id="re0", extension="arc")
-@blender_registry.register_archive_accessor(app_id="re1", extension="arc")
-@blender_registry.register_archive_accessor(app_id="re5", extension="arc")
-@blender_registry.register_archive_accessor(app_id="re6", extension="arc")
-@blender_registry.register_archive_accessor(app_id="rev1", extension="arc")
-@blender_registry.register_archive_accessor(app_id="rev2", extension="arc")
-@blender_registry.register_archive_accessor(app_id="dd", extension="arc")
-@blender_registry.register_archive_accessor(app_id="dmc4", extension="arc")
-def arc_accessor(vfile, context):
-    arc = ArcWrapper(vfile.root_vfile.absolute_path)
-    arc.app_id = vfile.app_id
-
-    path = vfile.relative_path_windows
-    path_no_ext = str(vfile.relative_path_windows_no_ext)
-    ext = path.suffix.replace(".", "")
-
-    # TODO: error handling, e.g. when file_path doesn't exist
-    try:
-        file_type = EXTENSION_TO_FILE_ID[ext]
-    except KeyError:
-        file_type = int(ext)
-    file_bytes = arc.get_file(path_no_ext, file_type)
-
-    return file_bytes
-
-
-class ArcWrapper:
-    PATH_SEPARATOR = "\\"
-
-    def __init__(self, file_path):
-        self.app_id = None
-        self.file_path = file_path
-        self.parsed = Arc.from_file(file_path)
-        self.parsed._read()
-
-    def get_file_entries_by_type(self, file_type):
-        filtered = []
-        for fe in self.parsed.file_entries:
-            if fe.file_type == file_type:
-                filtered.append(fe)
-        return filtered
-
-    def get_file_entries_by_extension(self, extension):
-        try:
-            file_type = EXTENSION_TO_FILE_ID[extension]
-        except KeyError:
-            raise RuntimeError(f"Extension {extension} unknown")
-        return self.get_file_entries_by_type(file_type)
-
-    def get_files_by_extension(self, extension):
-        try:
-            file_type = EXTENSION_TO_FILE_ID[extension]
-        except KeyError:
-            raise RuntimeError(f"Extension {extension} unknown")
-        files = []
-        for file_entry in self.get_file_entries_by_type(file_type):
-            t = (file_entry.file_path, self.get_file(
-                file_entry.file_path, file_type))
-            files.append(t)
-        return files
-
-    def get_file_entries(self):
-        file_entries = []
-        for fe in self.parsed.file_entries:
-            ext = FILE_ID_TO_EXTENSION.get(fe.file_type, fe.file_type)
-            fe.file_path_with_ext = f"{fe.file_path}.{ext}"
-            file_entries.append(fe)
-        return file_entries
-
-    def get_file(self, file_path, file_type):
-        file_ = None
-
-        for fe in self.parsed.file_entries:
-            if fe.file_path == file_path and fe.file_type == file_type:
-                try:
-                    if self.app_id == "dmc4":
-                        raise RuntimeError("The xcompression algoringm for DMC4 is not implemented yet.")
-                    else:
-                        file_ = zlib.decompress(fe.raw_data)
-                    break
-                except EOFError:
-                    print(
-                        f"Requested to read out of bounds. Offset: {fe.offset}")
-                    raise
-        return file_
-
-    def unpack(self, out_path):
-        arc_path = Path(self.file_path)
-        out_path = Path(out_path)
-
-        for fe in self.get_file_entries():
-            file_entry_path = PureWindowsPath(fe.file_path_with_ext)
-            out_file_path = out_path / arc_path.stem / file_entry_path
-            if not out_file_path.parent.exists():
-                os.makedirs(str(out_file_path.parent))
-            with open(str(out_file_path), "wb") as w:
-                data = self.get_file(fe.file_path, fe.file_type)
-                w.write(data)
+@blender_registry.register_fs_root_loader(app_id="re0", extension=None)
+@blender_registry.register_fs_root_loader(app_id="re1", extension=None)
+@blender_registry.register_fs_root_loader(app_id="re5", extension=None)
+@blender_registry.register_fs_root_loader(app_id="re6", extension=None)
+@blender_registry.register_fs_root_loader(app_id="rev1", extension=None)
+@blender_registry.register_fs_root_loader(app_id="rev2", extension=None)
+@blender_registry.register_fs_root_loader(app_id="dd", extension=None)
+@blender_registry.register_fs_root_loader(app_id="dmc4", extension=None)
+def game_fs_root_loader(absolute_path):
+    return MTFW_FS(absolute_path)
 
 
 def _sort_arc_entries(entries, vfile=True):
@@ -153,7 +63,7 @@ def _sort_arc_entries(entries, vfile=True):
 
 
 def _get_file_entry(vfile):
-    vf_data = vfile.data_bytes
+    vf_data = vfile.get_bytes()
     chunk = zlib.compress(vf_data)
     path = ntpath.normpath(vfile.relative_path)
     file_path = ntpath.splitext(path)[0]
@@ -178,7 +88,6 @@ def _serialize_arc(exported):
     header.ident = b"ARC\00"
     header.version = 7
     header.num_files = len(exported)
-    header._check()
     arc.header = header
     file_offset = header.num_files * 80 + -(header.num_files * 80) % 32768
 
@@ -192,12 +101,11 @@ def _serialize_arc(exported):
         file_entry.flags = 2
         file_entry.offset = file_offset
         file_entry.raw_data = fe.raw_data
-        file_entry._check()
         arc.file_entries.append(file_entry)
         file_offset += file_entry.zsize
 
     arc.padding = bytearray(32760 - (header.num_files * 80) % 32768)
-    arc._check()
+    check_recursive(arc)
 
     stream = KaitaiStream(io.BytesIO(bytearray(file_offset)))
     arc._write(stream)
@@ -323,7 +231,7 @@ def update_arc(filepath, vfiles, remove_unused_textures=False):
 
     # patch dictionary with imported files
     for vf in vf_sorted:
-        vf_data = vf.data_bytes
+        vf_data = vf.get_bytes()
         chunk = zlib.compress(vf_data)
         path = ntpath.normpath(vf.relative_path)
         file_path = ntpath.splitext(path)[0]
@@ -407,7 +315,7 @@ def find_and_replace_in_arc(filepath, vfile, file_name, add_new):
             if name == file_name and vfile.extension == extension:
                 show_message_box("File: {} was found and replaced in the archive".format(file_name))
                 found = True
-                vf_data = vfile.data_bytes
+                vf_data = vfile.get_bytes()
                 chunk = zlib.compress(vf_data)
                 fe.zsize = len(chunk)
                 fe.size = len(vf_data)