+++ CHANGELOG.md
@@ -23,10 +23,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Missed `FTransparencyDodgeMap` in `FTransparency` enum for RE6 maps
 - Import SBC for DMC4
 - Issue with SBC generation that broke high collision mesh (eff) and left holes in low collision mesh (scr)
+- Typo that set wrong name ("name") for the color attribute layer
+- Typo that brightened vertex colors on every import-export round trip
+- Vertex colors losing precision on every import-export round trip
+- Vertices export with multiple attributes, causing artifacts around UV seams
 
 ### Changed
 
 ### Removed
+
+- Split UV Seams tool, no longer needed now that export splits vertices automatically
 
 ## [0.5.0] - 2026-04-03